requirements.txt
#requirement txt
fastapi
uvicorn
pydub
Pillow
google-cloud-aiplatform